move to file by itself. Use inspect becasue autograph is using that under the hood
