+++ TrainingExtensions/tensorflow/test/python/eager/test_keras_utils.py
@@ -47,8 +47,8 @@ import tensorflow as tf
 from tensorflow.python.platform import gfile
 
 from aimet_tensorflow.keras.quantsim import QuantizationSimModel
-from aimet_tensorflow.keras.utils.common import replace_layer_in_functional_model, \
-    convert_h5_model_to_pb_model, prepare_model
+from aimet_tensorflow.keras.utils.common import convert_h5_model_to_pb_model
+from aimet_tensorflow.keras.utils.common import replace_layer_in_functional_model
 
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
@@ -57,10 +57,8 @@ def conv_functional():
     input_shape = (128, 28, 28, 1)
     inp = tf.keras.Input(shape=input_shape[1:])
     x = tf.keras.layers.Conv2D(32, kernel_size=(3, 3), activation="relu")(inp)
-    x = tf.keras.layers.Conv2DTranspose(
-        32, kernel_size=(3, 3), activation="relu")(x)
-    x = tf.keras.layers.DepthwiseConv2D(
-        depth_multiplier=1, kernel_size=(3, 3), activation='relu')(x)
+    x = tf.keras.layers.Conv2DTranspose(32, kernel_size=(3, 3), activation="relu")(x)
+    x = tf.keras.layers.DepthwiseConv2D(depth_multiplier=1, kernel_size=(3, 3), activation='relu')(x)
     x = tf.keras.layers.Flatten()(x)
     x = tf.keras.layers.Dropout(0.5, trainable=False)(x)
     x = tf.keras.layers.Dense(10, activation="softmax")(x)
@@ -78,8 +76,7 @@ class ConvTimesThree(tf.keras.layers.Layer):
                                            activation='relu',
                                            name='class_conv')
         self.conv_transpose = tf.keras.layers.Conv2DTranspose(64,
-                                                              kernel_size=(
-                                                                  3, 3),
+                                                              kernel_size=(3, 3),
                                                               activation='relu',
                                                               name='class_conv_transpose')
         self.depth_conv = tf.keras.layers.DepthwiseConv2D(depth_multiplier=1,
@@ -88,13 +85,16 @@ class ConvTimesThree(tf.keras.layers.Layer):
                                                           name='class_conv_depth')
 
     def call(self, x):
-        x = self.conv_transpose(self.conv(x))
-        return self.depth_conv(x)
+        x = self.conv(x)
+        x = self.conv_transpose(x)
+        x = self.depth_conv(x)
+        return x
+
 
 # See comment above ConvTimesThree Class
 def conv_sub_class():
     input_shape = (128, 28, 28, 1)
-    inp = tf.keras.Input(batch_shape=input_shape)
+    inp = tf.keras.Input(shape=input_shape[1:])
     x = ConvTimesThree()(inp)
     x = tf.keras.layers.Flatten()(x)
     x = tf.keras.layers.Dropout(0.5)(x, training=False)
@@ -103,84 +103,14 @@ def conv_sub_class():
     model = tf.keras.Model(inputs=inp, outputs=x, name='conv_classes')
     return model
 
-# Below models are based on Deep Learning with Python by Francois Chollet Second Edition (page 182 - 185)
-# Only Subclassing
-class CustomerTicketModel(tf.keras.Model):
-
-    def __init__(self, num_departments):
-        super().__init__()
-        self.concat_layer = tf.keras.layers.Concatenate()
-        self.mixing_layer = tf.keras.layers.Dense(64, activation="relu")
-        self.priority_scorer = tf.keras.layers.Dense(1, activation="sigmoid")
-        self.department_classifier = tf.keras.layers.Dense(num_departments, activation="softmax")
-
-    def call(self, inputs):
-        title = inputs["title"]
-        text_body = inputs["text_body"]
-        tags = inputs["tags"]
-
-        features = self.concat_layer([title, text_body, tags])
-        features = self.mixing_layer(features)
-        priority = self.priority_scorer(features)
-        department = self.department_classifier(features)
-        return priority, department
-
-# Functional model that includes subclassed layers
-class Classifier(tf.keras.Model):
-
-    def __init__(self, num_classes=4):
-        super().__init__()
-        if num_classes == 2:
-            num_units = 1
-            activation = "sigmoid"
-        else:
-            num_units = num_classes
-            activation = "softmax"
-        self.dense = tf.keras.layers.Dense(num_units, activation=activation)
-
-    def call(self, inputs):
-        return self.dense(inputs)
-
-def functional_model_with_subclassed_layers():
-    inputs = tf.keras.layers.Input(shape=(3,))
-    features = tf.keras.layers.Dense(64, activation="relu")(inputs)
-    outputs = Classifier(num_classes=10)(features)
-    model = tf.keras.Model(inputs=inputs, outputs=outputs)
-    return model
-
-
-# Subclass model that includes functional layers
-
-def subclass_model_with_functional_layers():
-    inputs = tf.keras.Input(shape=(64,))
-    outputs = tf.keras.layers.Dense(1, activation="sigmoid")(inputs)
-    binary_classifier = tf.keras.Model(inputs=inputs, outputs=outputs)
-
-    class MyModel(tf.keras.Model):
-
-        def __init__(self, num_classes=2):
-            super().__init__()
-            self.dense = tf.keras.layers.Dense(64, activation="relu")
-            self.classifier = binary_classifier
-
-        def call(self, inputs):
-            features = self.dense(inputs)
-            return self.classifier(features)
-
-    model = MyModel()
-    return model
-
-
 
 def test_replace_middle_layers():
     # Create model
     inp = tf.keras.layers.Input(shape=(2,))
     x = tf.keras.layers.Dense(units=1)(inp)
-    x = tf.keras.layers.Dense(
-        units=2, kernel_initializer=tf.keras.initializers.Constant(2.))(x)
+    x = tf.keras.layers.Dense(units=2, kernel_initializer=tf.keras.initializers.Constant(2.))(x)
     x = tf.keras.layers.Dense(units=3)(x)
-    model = tf.keras.Model(inputs=inp, outputs=x,
-                           name="replace_middle_layers_model")
+    model = tf.keras.Model(inputs=inp, outputs=x, name="replace_middle_layers_model")
 
     test_inp = np.array([[1, 2]])
     _ = model.predict(test_inp)
@@ -217,8 +147,7 @@ def test_replace_output_layer():
     inp = tf.keras.layers.Input(shape=(2,))
     x = tf.keras.layers.Dense(units=1)(inp)
     x = tf.keras.layers.Dense(units=2)(x)
-    model = tf.keras.Model(inputs=inp, outputs=x,
-                           name="replace_output_layer_model")
+    model = tf.keras.Model(inputs=inp, outputs=x, name="replace_output_layer_model")
 
     test_inp = np.array([[1, 2]])
     _ = model.predict(test_inp)
@@ -237,8 +166,7 @@ def test_replace_multi_input_layer():
     inp = tf.keras.layers.Input(shape=(2,))
     inp2 = tf.keras.layers.Input(shape=(2,))
     x = inp + inp2
-    model = tf.keras.Model(
-        inputs=[inp, inp2], outputs=x, name="replace_multi_input_layer_model")
+    model = tf.keras.Model(inputs=[inp, inp2], outputs=x, name="replace_multi_input_layer_model")
 
     test_inp = np.array([[1, 2]])
     test_inp2 = np.array([[2, 3]])
@@ -259,8 +187,7 @@ def test_replace_layer_with_multiple_children():
     x = tf.keras.layers.ReLU()(inp)
     out1 = tf.keras.layers.Dense(units=2)(x)
     out2 = tf.keras.layers.Dense(units=4)(x)
-    model = tf.keras.Model(inputs=inp, outputs=[
-                           out1, out2], name="replace_multiple_children_layer_model")
+    model = tf.keras.Model(inputs=inp, outputs=[out1, out2], name="replace_multiple_children_layer_model")
 
     test_inp = np.array([[1, 2]])
     _, _ = model.predict(test_inp)
@@ -283,12 +210,10 @@ def test_replace_layer_with_multiple_children():
 def test_replace_layer_in_internal_model():
     inp = tf.keras.layers.Input(shape=(2,))
     out = tf.keras.layers.PReLU(alpha_initializer='ones')(inp)
-    inner_model = tf.keras.Model(
-        inputs=inp, outputs=out, name="internal_model")
+    inner_model = tf.keras.Model(inputs=inp, outputs=out, name="internal_model")
 
     outer_model = tf.keras.Sequential()
-    outer_model.add(tf.keras.layers.PReLU(
-        alpha_initializer='ones', input_shape=(2,)))
+    outer_model.add(tf.keras.layers.PReLU(alpha_initializer='ones', input_shape=(2,)))
     outer_model.add(inner_model)
     outer_model.add(tf.keras.layers.PReLU(alpha_initializer='ones'))
 
@@ -378,11 +303,10 @@ def check_conversion_tensor_names(model, custom_objects=None):
         json.dump(quantsim_config, f)
 
     random_input_data = tf.random.normal(shape=(128, *model.input_shape[1:]))
-    # run forward pass to generate weights
-    _ = model.predict(random_input_data)
+    # run forward pass on dense to generate weights
+    _ = model(random_input_data)
 
-    sim = QuantizationSimModel(
-        model, quant_scheme='tf', config_file='./config.json')
+    sim = QuantizationSimModel(model, quant_scheme='tf', config_file='./config.json')
     sim.compute_encodings(lambda m, _: m.predict(random_input_data), None)
     # convert_h5_model_to_pb_model is called during export.
     sim.export('./tmp', model.name)
@@ -390,8 +314,7 @@ def check_conversion_tensor_names(model, custom_objects=None):
     # Get all encodings names (param_encodings and activation encodings) and put their respective keys
     # (which represent the weights names) into a set for fast checking.
     encodings = sim.get_encodings_dict()
-    encoding_weight_names = {
-        *encodings['param_encodings'].keys(), *encodings['activation_encodings'].keys()}
+    encoding_weight_names = {*encodings['param_encodings'].keys(), *encodings['activation_encodings'].keys()}
     original_weight_names = {
         weight_name.split(':')[0]
         for weight_name in encoding_weight_names
@@ -399,12 +322,10 @@ def check_conversion_tensor_names(model, custom_objects=None):
     }
 
     # Convert h5 model that was exported from QuantSim to a pb model to be used with encodings that were exported
-    converted_weight_names = get_converted_models_weight_names(
-        f'./tmp/{model.name}_converted.pb')
+    converted_weight_names = get_converted_models_weight_names(f'./tmp/{model.name}_converted.pb')
 
     # Check to see if all the original weight names can be found in the converted pb model
-    missing_weight_names = original_weight_names.difference(
-        converted_weight_names)
+    missing_weight_names = original_weight_names.difference(converted_weight_names)
     assert not missing_weight_names, f"Weight name(s): {missing_weight_names} are missing"
 
 
@@ -435,69 +356,3 @@ def test_convert_h5_to_pb_pretrained_keras():
     model = tf.keras.applications.ResNet50(weights="imagenet",
                                            input_shape=(224, 224, 3))
     check_conversion_tensor_names(model)
-
-
-def compare_weights(original_model, functional_model):
-    """
-    Helper function to compare the weights of two models. This function is used to test the conversion script.
-    :param original_model: the original model
-    :param functional_model: the model that was converted from the original model
-    """
-    original_weights = original_model.get_weights()
-    functional_weights = functional_model.get_weights()
-    for i in range(len(original_weights)):
-        assert np.array_equal(original_weights[i], functional_weights[i])
-
-
-def test_full_subclass_to_functional():
-    vocabulary_size = 10000
-    num_tags = 100
-    num_departments = 4
-    num_samples = 1280
-
-    title_data = np.random.randint(0, 2, size=(num_samples, vocabulary_size))
-    text_body_data = np.random.randint(0, 2, size=(num_samples, vocabulary_size))
-    tags_data = np.random.randint(0, 2, size=(num_samples, num_tags))
-
-
-    model = CustomerTicketModel(num_departments=num_departments)
-    _ = model({"title": title_data,
-               "text_body": text_body_data,
-               "tags": tags_data})
-    # Since this model is fully subclassed, specifically at the beginning, we call prepare model with
-    # the inputs to have Keras symoblic tensor fit the rest of the layers correctly.
-    functional_model = prepare_model(model,
-                                     [tf.keras.Input(shape=(num_samples, vocabulary_size,)),
-                                      tf.keras.Input(shape=(num_samples, vocabulary_size,)),
-                                      tf.keras.Input(shape=(num_samples, num_tags,))])
-    assert functional_model.count_params() == model.count_params()
-    compare_weights(model, functional_model)
-
-def test_functional_model_with_subclassed_layers_to_functional():
-    model = functional_model_with_subclassed_layers()
-    random_input = np.random.rand(32, 3)
-    _ = model(random_input)
-
-    functional_model = prepare_model(model)
-    assert functional_model.count_params() == model.count_params()
-    compare_weights(model, functional_model)
-
-def test_subclass_model_with_subclassed_layers_to_functional():
-    model = subclass_model_with_functional_layers()
-    input_shape = (32, 64)
-    random_input = np.random.rand(*input_shape)
-    _ = model(random_input)
-
-    functional_model = prepare_model(model, tf.keras.Input(shape=input_shape[1:]))
-    assert functional_model.count_params() == model.count_params()
-    compare_weights(model, functional_model)
-
-def test_conv_times_three_subclass_to_functional():
-    model = conv_sub_class()
-    input_shape = (32, 28, 28, 1)
-    random_input = np.random.rand(*input_shape)
-    _ = model(random_input)
-
-    functional_model = prepare_model(model)
-    assert functional_model.count_params() == model.count_params()
-    compare_weights(model, functional_model)
